Split VersionRange into one module per type
Extract Part, Version, Versions, Operator, Simple, and Complex into
their own modules under Type.VersionRange, matching the repo convention
of one type per module.
